+++ js/showGameHistory.js
@@ -1,0 +1,5 @@
+import { gameRecordHistory } from "./gameRecordSkeleton";
+
+export function showGameHistory() {
+    
+}