+++ changelog.md
@@ -1,7 +1,16 @@
 # 更新紀錄
 
-## Ver.1.0 (20230419)
+## Commit 1 (20230419)
 
 ### 完成
 * 完成產生數字、和基本的遊戲邏輯(比大小)
 * 學習使用 ESmodule 功能 還有 try...catch
+
+## Commit 2 (20230421)
+
+### 完成
+* GameRecordSkeleton.js, saveGameHistory, createNewGame 的變數、函數名稱更新
+* 改變GameRecordSkeleton中變數的結構
+
+### 待辦
+* 準備製作顯示遊戲過往紀錄的函數 showGameHistory